refactor: AdminPage 내부의 상품관리 로직을 ProductManager 컴포넌트로 관심사 분리
