+++ src/basic/__tests__/basic.test.tsx
@@ -2,8 +2,8 @@ import { act, fireEvent, render, renderHook, screen, within } from '@testing-lib
 import { useState } from 'react';
 import { describe, expect, test } from 'vitest';
 
-import { AdminPage } from '../../refactoring/components/AdminPage';
-import { CartPage } from '../../refactoring/components/CartPage';
+import { AdminPage } from '../../refactoring/components/domains/admin/AdminPage';
+import { CartPage } from '../../refactoring/components/domains/cart/CartPage';
 import { useCart, useCoupons, useProducts } from '../../refactoring/hooks';
 import * as cartUtils from '../../refactoring/models/cart';
