+++ src/refactoring/components/domains/admin/coupon/CreatedCouponList.tsx
@@ -1,4 +1,4 @@
-import type { Coupon } from '../../../types';
+import type { Coupon } from '../../../../../types';
 
 interface CreatedCouponListProps {
   coupons: Coupon[];
